# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Freelance Management API",
    "description": "Complete API documentation for freelance management system with products, orders, clients, and more",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/products/": {
      "get": {
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "example": "any"
                  },
                  "base_price": {
                    "example": "any"
                  },
                  "final_price": {
                    "example": "any"
                  },
                  "quantity": {
                    "example": "any"
                  },
                  "category_id": {
                    "example": "any"
                  },
                  "length": {
                    "example": "any"
                  },
                  "image_base64": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "example": "any"
                  },
                  "base_price": {
                    "example": "any"
                  },
                  "final_price": {
                    "example": "any"
                  },
                  "quantity": {
                    "example": "any"
                  },
                  "category_id": {
                    "example": "any"
                  },
                  "length": {
                    "example": "any"
                  },
                  "image_base64": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}